+++ src/platform_impl/windows/window_state.rs
@@ -110,6 +110,12 @@ bitflags! {
         /// Drop shadow for undecorated windows.
         const MARKER_UNDECORATED_SHADOW = 1 << 17;
 
+        /// When minmizing a maximized Windows, `WM_SIZE` which we use to update the `MAXIMIZED` bit
+        /// is fired with `wparam` set to `SIZE_MINIMIZED` and thus the `MAXIMIZED` bit will be unset
+        /// and when un-minimizing the window, the `MAXIMIZED` bit will still be unset and later on
+        /// in `apply_diff` when `new.to_window_styles()` is called, it will not add `WS_MAXIMIZE` window style.
+        const MARKER_WAS_MAXIMIZED = 1 << 18;
+
         const EXCLUSIVE_FULLSCREEN_OR_MASK = WindowFlags::ALWAYS_ON_TOP.bits;
     }
 }
@@ -357,13 +363,17 @@ impl WindowFlags {
         }
 
         if diff != WindowFlags::empty() {
-            let (style, style_ex) = new.to_window_styles();
+            let (mut style, style_ex) = new.to_window_styles();
 
             unsafe {
                 SendMessageW(window, *event_loop::SET_RETAIN_STATE_ON_SIZE_MSG_ID, 1, 0);
 
                 // This condition is necessary to avoid having an unrestorable window
                 if !new.contains(WindowFlags::MINIMIZED) {
+                    if self.contains(WindowFlags::MARKER_WAS_MAXIMIZED) {
+                        style |= WS_MAXIMIZE;
+                    }
+
                     SetWindowLongW(window, GWL_STYLE, style as i32);
                     SetWindowLongW(window, GWL_EXSTYLE, style_ex as i32);
                 }